requirements.txt
request
xmltodict
